requirements.txt
python-telegram-bot==20.0a2
vosk==0.3.42
aioredis==2.0.1
